fixed bug where elementToString didn't convert FD properly added optional validate flag to parseDA and parseTM to perform additional validation of the parsed data and throw if invalid updated unit tests to cover parseDA and parseTM changes

+++ src/util/elementToString.js
@@ -70,7 +70,7 @@ var dicomParser = (function (dicomParser) {
         }
         else if(vr == 'FD')
         {
-            textResult = multiElementToString(element.length / 8, dataSet.int32);
+            textResult = multiElementToString(element.length / 8, dataSet.double);
         }
         else if(vr == 'FL')
         {
